requirements.txt
opencv-python
pyzbar
pillow
pytesseract
xmltodict
numpy
